Refactor: move generic validators to base dir
