+++ packages/admin-panel-server/src/viz-builder/dashboardVisualisation/validators.ts
@@ -5,29 +5,6 @@
 
 import { yup } from '@tupaia/utils';
 
-import { configValidator } from './reportConfigValidator';
-
-export const baseVisualisationValidator = yup.object().shape({
-  presentation: yup.object().required(),
-  data: yup.object().required(),
-});
-
-export const baseVisualisationDataValidator = yup.object().shape({
-  fetch: yup.object().required(),
-});
-
-export const draftReportValidator = yup.object().shape({
-  code: yup.string().required('Requires "code" for the visualisation'),
-  config: configValidator,
-});
-
-export const legacyReportValidator = yup.object().shape({
-  code: yup.string().required('Requires "code" for the visualisation'),
-  dataBuilder: yup.string().required(),
-  config: yup.object().required(),
-  dataServices: yup.array().of(yup.object().shape({ isDataRegional: yup.boolean() })),
-});
-
 export const draftDashboardItemValidator = yup.object().shape({
   code: yup.string().required('Requires "code" for the visualisation'),
   config: yup.object().shape({ type: yup.string().required('Requires "type" in chart config') }),
