+++ packages/admin-panel-server/src/viz-builder/index.ts
@@ -5,3 +5,4 @@
 
 export * from './dashboardVisualisation';
 export * from './types';
+export * from './validators';
